Big reorganisation of introduction and theory chapters addressing Henning's comments
Henning's suggestion was to get to the main point of the thesis early on in the introduction. So I've moved most of the background information related to dark matter (evidence, types of searches, etc.) to the theory chapter instead. Tried to remove duplicate sentences (i.e., taken from the theory chapter to put in condensed intro) and integrate somewhat seamlessly into theory chapter